Fixes for the collada exporter. - resetting bones after each action export -> important if actions miss keyframes for certain bones - resetting object transformations. It can happen when animations are exported that objects are moved afterwards. - restore the same object selection as before the export - fixed image export for packed images. before the exporter has thrown an error because the image was not found from the resource folder. If file is packed the exporter saves it now directly from blender as a .png file
As far as I can tell there are no obvious errors currently. Tested with a few testscenes and imported them into godot.

+++ tools/export/blender25/io_scene_dae/export_dae.py
@@ -162,37 +162,55 @@ class DaeExporter:
 
 
 	def export_image(self,image):
-
 		if (image in self.image_cache):
 			return self.image_cache[image]
-
+			
 		imgpath = image.filepath
 		if (imgpath.find("//")==0 or imgpath.find("\\\\")==0):
 			#if relative, convert to absolute
 			imgpath = bpy.path.abspath(imgpath)
 
 		#path is absolute, now do something!
-
+		
 		if (self.config["use_copy_images"]):
 			#copy image
 			basedir = os.path.dirname(self.path)+"/images"
 			if (not os.path.isdir(basedir)):
 				os.makedirs(basedir)
-			dstfile=basedir+"/"+os.path.basename(imgpath)
-			if (not os.path.isfile(dstfile)):
-				shutil.copy(imgpath,dstfile)
-			imgpath="images/"+os.path.basename(imgpath)
+			
+			if os.path.isfile(imgpath):
+				dstfile=basedir+"/"+os.path.basename(imgpath)
+				
+				if (not os.path.isfile(dstfile)):
+					shutil.copy(imgpath,dstfile)
+					imgpath="images/"+os.path.basename(imgpath)
+			else:
+				### if file is not found save it as png file in the destination folder
+				img_tmp_path = image.filepath	
+				image.filepath = basedir+"/"+image.name+".png"
+					
+				dstfile=basedir+"/"+os.path.basename(image.filepath)
+				
+				if (not os.path.isfile(dstfile)):
+					
+					image.save()
+					imgpath="images/"+os.path.basename(image.filepath)
+				image.filepath = img_tmp_path
 
 		else:
 			#export relative, always, no one wants absolute paths.
 			try:
 				imgpath = os.path.relpath(imgpath,os.path.dirname(self.path)).replace("\\","/") # export unix compatible always
+				
 			except:
 				pass #fails sometimes, not sure why
-
-
+		
 
 		imgid = self.new_id("image")
+		
+		if (not os.path.isfile(imgpath)):
+			imgpath = imgpath="images/"+image.name+".png"
+		
 		self.writel(S_IMGS,1,'<image id="'+imgid+'" name="'+image.name+'">')
 		self.writel(S_IMGS,2,'<init_from>'+imgpath+'</init_from>"/>')
 		self.writel(S_IMGS,1,'</image>')
@@ -1173,6 +1191,7 @@ class DaeExporter:
 	def export_node(self,node,il):
 		if (not node in self.valid_nodes):
 			return
+		prev_node = bpy.context.scene.objects.active
 		bpy.context.scene.objects.active = node
 
 		self.writel(S_NODES,il,'<node id="'+self.validate_id(node.name)+'" name="'+node.name+'" type="NODE">')
@@ -1196,6 +1215,7 @@ class DaeExporter:
 
 		il-=1
 		self.writel(S_NODES,il,'</node>')
+		bpy.context.scene.objects.active = prev_node #make previous node active again
 
 	def is_node_valid(self,node):
 		if (not node.type in self.config["object_types"]):
@@ -1438,9 +1458,15 @@ class DaeExporter:
 		return tcn
 
 	def export_animations(self):
-
+		tmp_mat = []
+		for s in self.skeletons:
+			tmp_bone_mat = []
+			for bone in s.pose.bones:
+				tmp_bone_mat.append(Matrix(bone.matrix_basis))
+				bone.matrix_basis = Matrix()
+			tmp_mat.append([Matrix(s.matrix_local),tmp_bone_mat])
+			
 		self.writel(S_ANIM,0,'<library_animations>')
-
 
 
 		if (self.config["use_anim_action_all"] and len(self.skeletons)):
@@ -1473,13 +1499,18 @@ class DaeExporter:
 								bones.append(dp)
 
 				allowed_skeletons=[]
-				for y in self.skeletons:
+				for i,y in enumerate(self.skeletons):
 					if (y.animation_data):
 						for z in y.pose.bones:
 							if (z.bone.name in bones):
 								if (not y in allowed_skeletons):
 									allowed_skeletons.append(y)
 						y.animation_data.action=x;
+						
+						y.matrix_local = tmp_mat[i][0]
+						for j,bone in enumerate(s.pose.bones):
+							bone.matrix_basis = Matrix()
+							
 
 				print("allowed skeletons "+str(allowed_skeletons))
 
@@ -1498,16 +1529,20 @@ class DaeExporter:
 
 			self.writel(S_ANIM_CLIPS,0,'</library_animation_clips>')
 
-			for s in self.skeletons:
+			for i,s in enumerate(self.skeletons):
 				if (s.animation_data==None):
 					continue
 				if s in cached_actions:
 					s.animation_data.action = bpy.data.actions[cached_actions[s]]
 				else:
 					s.animation_data.action = None
+					for j,bone in enumerate(s.pose.bones):
+						bone.matrix_basis = tmp_mat[i][1][j]
 		else:
 			self.export_animation(self.scene.frame_start,self.scene.frame_end)
-
+		
+			
+		
 		self.writel(S_ANIM,0,'</library_animations>')
 
 	def export(self):
